+++ scripts/cleanup.rs
@@ -1,3 +1,4 @@
+#!/usr/bin/env rust-script
 //! Cleans up all files and folders that were produced by test runs.
 //!
 //! ```cargo