+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.25.6] - 2023-07-28
+
+Add methods
+`LoggerHandle::adapt_duplication_to_stderr` and  `LoggerHandle::adapt_duplication_to_stdout`
+(realizes issue #142).
+
+Extend docu on providing custom format.
+
+Use rust-script instead of cargo-script for qualification scripts.
+
+Update dependencies.
+
 ## [0.25.5] - 2023-05-25
 
 Use display (rather than debug) formatting for thread names